fix controller

+++ sdwan-controller/src/main/java/io/jaspercloud/sdwan/server/service/impl/RouteServiceImpl.java
@@ -6,11 +6,12 @@ import io.jaspercloud.sdwan.exception.ProcessException;
 import io.jaspercloud.sdwan.server.controller.request.EditRouteRequest;
 import io.jaspercloud.sdwan.server.controller.response.PageResponse;
 import io.jaspercloud.sdwan.server.entity.Route;
+import io.jaspercloud.sdwan.server.entity.RouteNodeItem;
 import io.jaspercloud.sdwan.server.repository.RouteNodeItemRepository;
 import io.jaspercloud.sdwan.server.repository.RouteRepository;
 import io.jaspercloud.sdwan.server.repository.po.RouteNodeItemPO;
 import io.jaspercloud.sdwan.server.repository.po.RoutePO;
-import io.jaspercloud.sdwan.server.service.NodeConfigService;
+import io.jaspercloud.sdwan.server.service.GroupConfigService;
 import io.jaspercloud.sdwan.server.service.RouteService;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
@@ -18,12 +19,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
 @Transactional
 public class RouteServiceImpl implements RouteService {
-
 
     @Resource
     private RouteRepository routeRepository;
@@ -32,46 +33,42 @@ public class RouteServiceImpl implements RouteService {
     private RouteNodeItemRepository routeNodeItemRepository;
 
     @Resource
-    private NodeConfigService nodeConfigService;
+    private GroupConfigService groupConfigService;
 
     @Override
     public void add(EditRouteRequest request) {
         RoutePO route = BeanUtil.toBean(request, RoutePO.class);
         route.setId(null);
         route.insert();
-        for (Long nodeId : request.getNodeIdList()) {
-            boolean exists = nodeConfigService.existsNode(nodeId);
-            if (!exists) {
-                throw new ProcessException("not found node");
+        if (CollectionUtil.isNotEmpty(request.getNodeIdList())) {
+            for (Long nodeId : request.getNodeIdList()) {
+                RouteNodeItemPO routeNodeItem = new RouteNodeItemPO();
+                routeNodeItem.setRouteId(route.getId());
+                routeNodeItem.setNodeId(nodeId);
+                routeNodeItem.insert();
             }
-            RouteNodeItemPO routeNodeItem = new RouteNodeItemPO();
-            routeNodeItem.setRouteId(route.getId());
-            routeNodeItem.setNodeId(nodeId);
-            routeNodeItem.insert();
         }
     }
 
     @Override
     public void edit(EditRouteRequest request) {
         Route route = BeanUtil.toBean(request, Route.class);
-        routeNodeItemRepository.delete(routeNodeItemRepository.lambdaQuery()
-                .eq(RouteNodeItemPO::getRouteId, request.getId()));
-        for (Long nodeId : request.getNodeIdList()) {
-            boolean exists = nodeConfigService.existsNode(nodeId);
-            if (!exists) {
-                throw new ProcessException("not found node");
-            }
-            RouteNodeItemPO routeNodeItem = new RouteNodeItemPO();
-            routeNodeItem.setRouteId(route.getId());
-            routeNodeItem.setNodeId(nodeId);
-            routeNodeItem.insert();
-        }
         routeRepository.updateById(route);
+        if (CollectionUtil.isNotEmpty(request.getNodeIdList())) {
+            routeNodeItemRepository.delete(routeNodeItemRepository.lambdaQuery()
+                    .eq(RouteNodeItemPO::getRouteId, request.getId()));
+            for (Long nodeId : request.getNodeIdList()) {
+                RouteNodeItemPO routeNodeItem = new RouteNodeItemPO();
+                routeNodeItem.setRouteId(route.getId());
+                routeNodeItem.setNodeId(nodeId);
+                routeNodeItem.insert();
+            }
+        }
     }
 
     @Override
     public void del(EditRouteRequest request) {
-        if (nodeConfigService.usedRoute(request.getId())) {
+        if (groupConfigService.usedRoute(request.getId())) {
             throw new ProcessException("node used");
         }
         routeRepository.deleteById(request.getId());
@@ -103,6 +100,20 @@ public class RouteServiceImpl implements RouteService {
                 .stream().map(e -> e.getNodeId()).collect(Collectors.toList());
         route.setNodeIdList(collect);
         return route;
+    }
+
+    @Override
+    public List<Route> queryDetailByIdList(List<Long> idList) {
+        List<Route> routeList = routeRepository.selectBatchIds(idList);
+        Map<Long, List<RouteNodeItem>> map = routeNodeItemRepository.list(routeNodeItemRepository.lambdaQuery()
+                        .in(RouteNodeItemPO::getRouteId, routeList.stream().map(e -> e.getId()).collect(Collectors.toList())))
+                .stream().collect(Collectors.groupingBy(e -> e.getRouteId()));
+        routeList.forEach(route -> {
+            List<Long> collect = map.getOrDefault(route.getId(), Collections.emptyList())
+                    .stream().map(e -> e.getNodeId()).collect(Collectors.toList());
+            route.setNodeIdList(collect);
+        });
+        return routeList;
     }
 
     @Override
